Revert player gains to dynamic method
