+++ server/src/api/services/internal/delta.service.ts
@@ -4,12 +4,31 @@ import { Pagination } from 'src/types';
 import { Delta, InitialValues, Player, Snapshot } from '../../../database/models';
 import { ALL_METRICS, PERIODS, PLAYER_BUILDS, PLAYER_TYPES } from '../../constants';
 import { BadRequestError } from '../../errors';
-import { getPlayerPeriodDeltasLegacy, getSeconds } from '../../util/legacy';
 import { getMeasure, getRankKey, getValueKey, isSkill } from '../../util/metrics';
 import { buildQuery } from '../../util/query';
 import * as snapshotService from './snapshot.service';
 
 const DELTA_INDICATORS = ['value', 'rank', 'efficiency'];
+
+export const DAY_IN_SECONDS = 86400;
+export const WEEK_IN_SECONDS = 604800;
+export const MONTH_IN_SECONDS = 2678400; // month = 31 days (like CML)
+export const YEAR_IN_SECONDS = 31556926;
+
+export function getSeconds(period) {
+  switch (period) {
+    case 'day':
+      return DAY_IN_SECONDS;
+    case 'week':
+      return WEEK_IN_SECONDS;
+    case 'month':
+      return MONTH_IN_SECONDS;
+    case 'year':
+      return YEAR_IN_SECONDS;
+    default:
+      return -1;
+  }
+}
 
 async function syncDeltas(playerId: number, period: string, latest: Snapshot, initial: InitialValues) {
   // prettier-ignore
@@ -102,7 +121,12 @@ async function syncInitialValues(playerId: number, latest: Snapshot) {
 /**
  * Get all the player deltas (gains) for a specific time period.
  */
-async function getPlayerPeriodDeltas(playerId: number, period: string) {
+async function getPlayerPeriodDeltas(
+  playerId: number,
+  period: string,
+  latest?: Snapshot,
+  initial?: InitialValues
+) {
   if (!playerId) {
     throw new BadRequestError('Invalid player id.');
   }
@@ -111,54 +135,21 @@ async function getPlayerPeriodDeltas(playerId: number, period: string) {
     throw new BadRequestError(`Invalid period: ${period}.`);
   }
 
-  const deltas = await Delta.findAll({ where: { playerId, period } });
+  const periodStartDate = Date.now() - getSeconds(period) * 1000;
+  const initialValues = initial || (await InitialValues.findOne({ where: { playerId } }));
+  const latestSnapshot = latest || (await snapshotService.findLatest(playerId));
 
-  // TODO: remove in a few weeks
-  // If the player hasn't been updated since this update came out
-  // use the legacy deltas calculations.
-  if (!deltas || deltas.length < DELTA_INDICATORS.length) {
-    const legacyGains = await getPlayerPeriodDeltasLegacy(playerId, period);
-    return legacyGains;
+  const startSnapshot = await snapshotService.findFirstSince(playerId, periodStartDate);
+
+  if (!startSnapshot || !latestSnapshot) {
+    return { period, startsAt: null, endsAt: null, data: emptyDiff() };
   }
-
-  const latest = await snapshotService.findLatest(playerId);
-
-  const rankDeltas = deltas.find(d => d.indicator === 'rank');
-  const valueDeltas = deltas.find(d => d.indicator === 'value');
-
-  const formattedData = Object.fromEntries(
-    ALL_METRICS.map(metric => {
-      const rankEnd = parseInt(latest[getRankKey(metric)]);
-      const rankGained = parseInt(rankDeltas[metric] || 0);
-      const rankStart = rankEnd - rankGained;
-
-      const valueEnd = parseInt(latest[getValueKey(metric)]);
-      const valueGained = parseInt(valueDeltas[metric] || 0);
-      const valueStart = valueEnd - valueGained;
-
-      return [
-        metric,
-        {
-          rank: {
-            start: rankStart,
-            end: rankEnd,
-            gained: rankGained
-          },
-          [getMeasure(metric)]: {
-            start: valueStart,
-            end: valueEnd,
-            gained: valueGained
-          }
-        }
-      ];
-    })
-  );
 
   return {
     period,
-    startsAt: deltas[0].startedAt,
-    endsAt: deltas[0].endedAt,
-    data: formattedData
+    startsAt: startSnapshot.createdAt,
+    endsAt: latestSnapshot.createdAt,
+    data: diff(startSnapshot, latestSnapshot, initialValues)
   };
 }
 
@@ -166,9 +157,12 @@ async function getPlayerPeriodDeltas(playerId: number, period: string) {
  * Gets the all the player deltas (gains), for every period.
  */
 async function getPlayerDeltas(playerId: number) {
+  const initial = await InitialValues.findOne({ where: { playerId } });
+  const latest = await snapshotService.findLatest(playerId);
+
   const periodDeltas = await Promise.all(
     PERIODS.map(async period => {
-      const deltas = await getPlayerPeriodDeltas(playerId, period);
+      const deltas = await getPlayerPeriodDeltas(playerId, period, latest, initial);
       return { period, deltas };
     })
   );
@@ -241,6 +235,61 @@ async function getGroupLeaderboard(
     gained: parseInt(d[metric]),
     player: d.player
   }));
+}
+
+/**
+ * Calculate the difference between two snapshots,
+ * taking untracked values into consideration. (via initial values)
+ */
+function diff(start: Snapshot, end: Snapshot, initial: InitialValues) {
+  const diffObj = {};
+
+  ALL_METRICS.forEach(metric => {
+    const rankKey = getRankKey(metric);
+    const valueKey = getValueKey(metric);
+
+    const initialRank = initial ? initial[rankKey] : -1;
+    const initialValue = initial ? initial[valueKey] : -1;
+
+    const endValue = parseInt(end[valueKey], 10);
+    const endRank = end[rankKey];
+
+    const startValue = parseInt(start[valueKey] === -1 ? initialValue : start[valueKey], 10);
+    const startRank = start[rankKey] === -1 && !isSkill(metric) ? initialRank : start[rankKey];
+
+    // Do not use initial ranks for skill, to prevent -1 ranks
+    // introduced by https://github.com/wise-old-man/wise-old-man/pull/93 from creating crazy diffs
+    const gainedRank = isSkill(metric) && start[rankKey] === -1 ? 0 : endRank - startRank;
+    const gainedValue = endValue - startValue;
+
+    diffObj[metric] = {
+      rank: {
+        start: startRank,
+        end: endRank,
+        gained: gainedRank
+      },
+      [getMeasure(metric)]: {
+        start: startValue,
+        end: endValue,
+        gained: gainedValue
+      }
+    };
+  });
+
+  return diffObj;
+}
+
+function emptyDiff() {
+  const diffObj = {};
+
+  ALL_METRICS.forEach(metric => {
+    diffObj[metric] = {
+      rank: { start: 0, end: 0, gained: 0 },
+      [getMeasure(metric)]: { start: 0, end: 0, gained: 0 }
+    };
+  });
+
+  return diffObj;
 }
 
 export {
